+++ raiden-dapp/src/components/account/AccountContent.vue
@@ -2,32 +2,33 @@
   <div class="account-content">
     <div v-if="!loading && defaultAccount">
       <v-row class="account-content__account-details" no-gutters>
-        <v-col cols="12">
-          <div class="account-content__account-details--title">
-            {{ $t('account-content.account-details') }}
-          </div>
-        </v-col>
-      </v-row>
-      <v-row no-gutters>
         <v-col cols="2">
-          <span class="account-content__account-details--address">
+          <span class="account-content__account-details__address">
             {{ $t('account-content.address') }}
           </span>
         </v-col>
         <v-col cols="10">
-          <span class="account-content__account-details--address">
-            <address-display :address="defaultAccount" full-address />
+          <span class="account-content__account-details__address">
+            <address-display
+              class="account-content__account-details__address__desktop"
+              :address="defaultAccount"
+              full-address
+            />
+            <address-display
+              class="account-content__account-details__address__mobile"
+              :address="defaultAccount"
+            />
           </span>
         </v-col>
       </v-row>
       <v-row class="account-content__account-details__eth" no-gutters>
         <v-col cols="2">
-          <span class="account-content__account-details__eth--currency">
+          <span class="account-content__account-details__eth__currency">
             {{ $t('account-content.currency') }}
           </span>
         </v-col>
         <v-col cols="10">
-          <span class="account-content__account-details__eth--balance">
+          <span class="account-content__account-details__eth__balance">
             {{ balance | decimals }}
           </span>
         </v-col>
@@ -45,8 +46,7 @@
             max-width="40px"
             height="36px"
             contain
-          >
-          </v-img>
+          ></v-img>
         </div>
         <v-list-item-content>
           <v-list-item-title>{{ menuItem.title }}</v-list-item-title>
@@ -187,6 +187,7 @@ export default class AccountContent extends Mixins(NavigationMixin) {
 </script>
 
 <style scoped lang="scss">
+@import '../../scss/mixins';
 @import '../../scss/colors';
 
 .account-content {
@@ -203,10 +204,32 @@ export default class AccountContent extends Mixins(NavigationMixin) {
 
     &__address {
       font-size: 16px;
+
+      &__desktop {
+        margin-top: 2px;
+        @include respond-to(handhelds) {
+          display: none;
+        }
+      }
+
+      &__mobile {
+        display: none;
+        margin-top: 2px;
+        @include respond-to(handhelds) {
+          display: flex;
+          margin-left: 30px;
+        }
+      }
     }
 
     &__eth {
       margin-bottom: 66px;
+
+      &__balance {
+        @include respond-to(handhelds) {
+          margin-left: 30px;
+        }
+      }
 
       &__currency,
       &__balance {