+++ raiden-dapp/tests/unit/components/account/account-content.spec.ts
@@ -38,35 +38,33 @@ describe('AccountContent.vue', () => {
     await wrapper.vm.$nextTick();
   });
 
-  test('displays account details title', async () => {
-    const accountDetailsTitle = wrapper.find(
-      '.account-content__account-details--title'
-    );
-
-    expect(accountDetailsTitle.text()).toBe('account-content.account-details');
-  });
-
   test('displays address', async () => {
     store.commit('account', '0x31aA9D3E2bd38d22CA3Ae9be7aae1D518fe46043');
     await wrapper.vm.$nextTick();
     const addressTitle = wrapper
-      .findAll('.account-content__account-details--address')
+      .findAll('.account-content__account-details__address')
       .at(0);
-    const address = wrapper
-      .findAll('.account-content__account-details--address')
-      .at(1);
+    const addressDesktop = wrapper.find(
+      '.account-content__account-details__address__desktop'
+    );
+    const addressMobile = wrapper.find(
+      '.account-content__account-details__address__mobile'
+    );
 
     expect(addressTitle.text()).toBe('account-content.address');
-    expect(address.text()).toBe('0x31aA9D3E2bd38d22CA3Ae9be7aae1D518fe46043');
+    expect(addressDesktop.text()).toBe(
+      '0x31aA9D3E2bd38d22CA3Ae9be7aae1D518fe46043'
+    );
+    expect(addressMobile.text()).toBe('0x31...6043');
   });
 
   test('displays eth', async () => {
     store.commit('balance', '12.0');
     await wrapper.vm.$nextTick();
     const ethTitle = wrapper.find(
-      '.account-content__account-details__eth--currency'
+      '.account-content__account-details__eth__currency'
     );
-    const eth = wrapper.find('.account-content__account-details__eth--balance');
+    const eth = wrapper.find('.account-content__account-details__eth__balance');
 
     expect(ethTitle.text()).toBe('account-content.currency');
     expect(eth.text()).toBe('12.000');